requirements.txt
discord.py
yt-dlp
PyNaCl
google-genai
